requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
_pytorch_select=0.1=cpu_0
absl-py=1.2.0=pypi_0
alsa-lib=1.2.3=h516909a_0
blas=1.0=mkl
bottleneck=1.3.2=py39hdd57654_1
brotli=1.0.9=he6710b0_2
bzip2=1.0.8=h7b6447c_0
c-ares=1.18.1=h7f8727e_0
ca-certificates=2021.10.8=ha878542_0
cachetools=5.2.0=pypi_0
cachetoolsutils=6.0=pypi_0
cairo=1.16.0=hf32fb01_1
certifi=2021.10.8=py39hf3d152e_1
cffi=1.15.0=py39hd667e15_1
charset-normalizer=2.1.1=pypi_0
cloudpickle=1.6.0=pyhd3eb1b0_0
colorama=0.4.6=pypi_0
cycler=0.11.0=pyhd3eb1b0_0
cython=0.29.27=pypi_0
cyvlfeat=0.6.0=py39h075484b_2
dbus=1.13.18=hb2f20db_0
expat=2.4.4=h295c915_0
farama-notifications=0.0.4=pypi_0
fasteners=0.17.3=pypi_0
ffmpeg=4.3.2=hca11adc_0
fontconfig=2.13.1=h6c09931_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.11.0=h70c0345_0
future=0.18.2=py39h06a4308_1
gettext=0.21.0=hf68c758_0
giflib=5.2.1=h7b6447c_0
glfw=2.5.0=pypi_0
glib=2.68.3=h9c3ff4c_0
glib-tools=2.68.3=h9c3ff4c_0
gmp=6.2.1=h2531618_2
gnutls=3.6.15=he1e5248_0
google-auth=2.12.0=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
graphite2=1.3.14=h23475e2_0
grpcio=1.49.1=pypi_0
gst-plugins-base=1.18.4=hf529b03_2
gstreamer=1.18.4=h76c114f_2
gym=0.19.0=py39hef51801_0
gymnasium=0.29.0=pypi_0
harfbuzz=2.8.2=h83ec7ef_0
hdf5=1.10.6=nompi_h6a2412b_1114
icu=68.1=h2531618_0
idna=3.4=pypi_0
imageio=2.16.0=pypi_0
importlib-metadata=5.0.0=pypi_0
intel-openmp=2019.4=243
jasper=1.900.1=hd497a04_4
jbig=2.1=hdba287a_0
joblib=1.2.0=pypi_0
jpeg=9d=h7f8727e_0
kiwisolver=1.3.2=py39h295c915_0
krb5=1.19.2=hac12032_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.35.1=h7274673_9
lerc=2.2.1=h2531618_0
libblas=3.9.0=1_h86c2bf4_netlib
libcblas=3.9.0=5_h92ddd45_netlib
libclang=11.1.0=default_ha53f305_1
libcurl=7.80.0=h0b77cf5_0
libdeflate=1.7=h27cfd23_5
libedit=3.1.20210910=h7f8727e_0
libev=4.33=h7f8727e_1
libevent=2.1.10=hcdb4288_3
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h5101ec6_17
libgfortran-ng=9.3.0=ha5ec8a7_17
libgfortran5=9.3.0=ha5ec8a7_17
libglib=2.68.3=h3e27bee_0
libgomp=9.3.0=h5101ec6_17
libiconv=1.16=h516909a_0
libidn2=2.3.2=h7f8727e_0
liblapack=3.9.0=5_h92ddd45_netlib
liblapacke=3.9.0=5_h92ddd45_netlib
libllvm11=11.1.0=h3826bc1_0
libmklml=2019.0.5=h06a4308_0
libnghttp2=1.46.0=hce63b2e_0
libogg=1.3.5=h27cfd23_1
libopencv=4.5.0=py39_2
libopus=1.3.1=h7b6447c_0
libpng=1.6.37=hbc83047_0
libpq=13.3=hd57d9b9_0
libssh2=1.9.0=h1ba5d50_1
libstdcxx-ng=9.3.0=hd4cf53a_17
libtasn1=4.16.0=h27cfd23_0
libtiff=4.3.0=hf544144_1
libtmux=0.22.2=pypi_0
libunistring=0.9.10=h27cfd23_0
libuuid=1.0.3=h7f8727e_2
libvorbis=1.3.7=h7b6447c_0
libwebp=1.2.2=h55f646e_0
libwebp-base=1.2.2=h7f8727e_0
libxcb=1.14=h7b6447c_0
libxkbcommon=1.0.3=he3ba5ed_0
libxml2=2.9.12=h72842e0_0
lz4-c=1.9.3=h295c915_1
markdown=3.4.1=pypi_0
markupsafe=2.1.1=pypi_0
matplotlib-base=3.5.1=py39ha18d171_0
menpo=0.10.0=py39ha90915f_3
mkl=2020.2=256
mkl-service=2.3.0=py39he8ac12f_0
mkl_fft=1.3.0=py39h54f3939_0
mkl_random=1.0.2=py39h63df603_0
mujoco-py=2.1.2.14=pypi_0
munkres=1.1.4=py_0
mysql-common=8.0.25=ha770c72_2
mysql-libs=8.0.25=hfa10184_2
ncurses=6.3=h7f8727e_2
nettle=3.7.3=hbbd107a_1
networkx=2.7.1=pypi_0
ninja=1.10.2=py39hd09550d_3
nspr=4.33=h295c915_0
nss=3.74=h0370c37_0
numexpr=2.7.3=py39hb2eb853_0
numpy=1.19.0=pypi_0
oauthlib=3.2.1=pypi_0
opencv=4.5.0=py39_2
opencv-python=4.7.0.72=pypi_0
openh264=2.1.1=h4ff587b_0
openssl=1.1.1m=h7f8727e_0
osmesa=12.2.2.dev=0
packaging=21.3=pyhd3eb1b0_0
pandas=1.4.1=py39h295c915_0
patchelf=0.13=h295c915_0
pcre=8.45=h295c915_0
pillow=9.0.1=pypi_0
pip=21.2.4=py39h06a4308_0
pixman=0.40.0=h7f8727e_1
protobuf=3.19.6=pypi_0
py-opencv=4.5.0=py39h9cd51e4_2
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pyglet=1.5.16=py39hf3d152e_1
pyparsing=3.0.4=pyhd3eb1b0_0
python=3.9.7=h12debd9_1
python-dateutil=2.8.2=pyhd3eb1b0_0
python_abi=3.9=2_cp39
pytz=2021.3=pyhd3eb1b0_0
pyyaml=6.0.1=pypi_0
qt=5.12.9=hda022c4_4
readline=8.1.2=h7f8727e_1
requests=2.28.1=pypi_0
requests-oauthlib=1.3.1=pypi_0
rsa=4.9=pypi_0
rtree=0.9.7=pypi_0
scikit-learn=1.1.2=pypi_0
scipy=1.5.3=py39hee8e79c_0
setuptools=58.0.4=py39h06a4308_0
six=1.16.0=pyhd3eb1b0_1
sklearn=0.0=pypi_0
sqlite=3.37.2=hc218d9a_0
stable-baselines3=1.4.0=pyhd8ed1ab_0
tensorboard=2.10.1=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
termplotlib=0.3.9=pypi_0
threadpoolctl=3.1.0=pypi_0
tk=8.6.11=h1ccaba5_0
tmuxp=1.28.2=pypi_0
torch=1.12.0=pypi_0
torchvision=0.13.0=pypi_0
tqdm=4.64.1=pypi_0
trimesh=3.10.2=pypi_0
typing=3.7.4.3=pypi_0
typing-extensions=4.7.1=pypi_0
tzdata=2021e=hda174b7_0
urllib3=1.26.12=pypi_0
vlfeat=0.9.20=h14c3975_1002
webcolors=1.13=pypi_0
werkzeug=2.2.2=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
x264=1!161.3030=h7f98852_1
xz=5.2.5=h7b6447c_0
zipp=3.9.0=pypi_0
zlib=1.2.11=h7f8727e_4
zstd=1.5.0=ha4553b6_1
